moved PrettyWriter from core to compiler-core
