+++ source/compiler-core/slang-pretty-writer.cpp
@@ -1,5 +1,5 @@
 #include "slang-pretty-writer.h"
-#include "slang-string-escape-util.h"
+#include "../core/slang-string-escape-util.h"
 
 namespace Slang
 {
