+++ tools/slang-reflection-test/slang-reflection-test-main.cpp
@@ -4,7 +4,7 @@
 #include "../../source/core/slang-string-escape-util.h"
 #include "../../source/core/slang-string-util.h"
 #include "../../source/core/slang-test-tool-util.h"
-#include "../../source/core/slang-pretty-writer.h"
+#include "../../source/compiler-core/slang-pretty-writer.h"
 #include "../../source/slang/slang-reflection-json.h"
 #include "slang-com-helper.h"
 #include "slang.h"
